+++ example/storybook-nativewind/src/components/Popover/Popover.tsx
@@ -1,24 +1,5 @@
 import React, { useState } from 'react';
 import {
-  Text,
-  Button,
-  ButtonText,
-  ButtonGroup,
-  CloseIcon,
-  Icon,
-  Box,
-  Center,
-  Heading,
-  Pressable,
-  HStack,
-  VStack,
-  Avatar,
-  AvatarFallbackText,
-  CircleIcon,
-  AddIcon,
-  PopoverArrow,
-} from '../../core-components/themed';
-import {
   Popover,
   PopoverBackdrop,
   PopoverContent,
@@ -26,23 +7,29 @@ import {
   PopoverBody,
   PopoverFooter,
   PopoverCloseButton,
+  PopoverArrow,
 } from '@/components/ui/popover';
+import { Button, ButtonText } from '@/components/ui/button';
+import { Text } from '@/components/ui/text';
+import { Heading } from '@/components/ui/heading';
+import { Box } from '@/components/ui/box';
+import { X } from 'lucide-react-native';
 
-import { PhoneIcon, Clock3Icon, MailIcon } from 'lucide-react-native';
-function PopoverBasic(props: any) {
-  const [isOpen, setIsOpen] = React.useState(false);
-  const handleOpen = () => {
-    setIsOpen(true);
-  };
-  const handleClose = () => {
-    setIsOpen(false);
-  };
+const PopoverBasic = (props: any) => {
+  const [showPopover, setShowPopover] = React.useState(false);
   return (
     <Popover
-      isOpen={isOpen}
-      onClose={handleClose}
-      onOpen={handleOpen}
       {...props}
+      placement="right"
+      shouldFlip
+      isOpen={showPopover}
+      onOpen={() => {
+        setShowPopover(true);
+      }}
+      onClose={() => {
+        setShowPopover(false);
+      }}
+      // eslint-disable-next-line react/no-unstable-nested-components
       trigger={(triggerProps) => {
         return (
           <Button {...triggerProps}>
@@ -53,90 +40,47 @@ function PopoverBasic(props: any) {
     >
       <PopoverBackdrop />
       <PopoverContent>
+        <PopoverArrow />
         <PopoverHeader>
-          <Heading size="lg">Welcome!</Heading>
+          <Heading>Welcome!</Heading>
           <PopoverCloseButton>
-            <Icon as={CloseIcon} />
+            <X
+              size={20}
+              className="stroke-background-400 group-[:hover]/popover-close-button:stroke-background-700 group-[:active]/popover-close-button:stroke-background-900 group-[:focus-visible]/popover-close-button:stroke-background-900"
+            />
           </PopoverCloseButton>
         </PopoverHeader>
         <PopoverBody>
-          <Text size="sm">
+          <Text>
             Join the product tour and start creating your own checklist. Are you
             ready to jump in?
           </Text>
         </PopoverBody>
         <PopoverFooter>
-          <Text size="xs" flex={1}>
+          <Text size="xs" className="flex-1">
             Step 2 of 3
           </Text>
-          <ButtonGroup space="md">
-            <Button variant="outline" action="secondary" onPress={handleClose}>
-              <ButtonText>Back</ButtonText>
-            </Button>
-            <Button onPress={handleClose}>
-              <ButtonText>Next</ButtonText>
-            </Button>
-          </ButtonGroup>
+          {/* @ts-ignore */}
+          <Button
+            variant="outline"
+            action="secondary"
+            className="mr-3"
+            onPress={() => {
+              setShowPopover(false);
+            }}
+          >
+            <ButtonText>Back</ButtonText>
+          </Button>
+          <Button
+            onPress={() => {
+              setShowPopover(false);
+            }}
+          >
+            <ButtonText>Next</ButtonText>
+          </Button>
         </PopoverFooter>
       </PopoverContent>
     </Popover>
-  );
-}
-
-const FigmaPopoverStory = ({
-  showPopover: _showPopoverProp = true,
-  _placement = 'bottom',
-  ...props
-}: any) => {
-  return (
-    <Box w="$full" h="$full">
-      <Popover
-        {...props}
-        _experimentalOverlay={true}
-        offset={30}
-        isOpen={true}
-        placement="top"
-        // eslint-disable-next-line react/no-unstable-nested-components
-        trigger={(triggerProps) => {
-          return (
-            <Box w={1200} h={500} pt={300} pb={50} alignItems="center">
-              <Button {...triggerProps}>
-                <ButtonText>Popover</ButtonText>
-              </Button>
-            </Box>
-          );
-        }}
-      >
-        <PopoverContent>
-          <PopoverHeader>
-            <Heading>Welcome!</Heading>
-            <PopoverCloseButton>
-              <Icon as={CloseIcon} />
-            </PopoverCloseButton>
-          </PopoverHeader>
-          <PopoverBody>
-            <Text>
-              Join the product tour and start creating your own checklist. Are
-              you ready to jump in?
-            </Text>
-          </PopoverBody>
-          <PopoverFooter>
-            <Text size="xs" flex={1}>
-              Step 2 of 3
-            </Text>
-            {/* @ts-ignore */}
-            <ButtonGroup space="md">
-              <Button variant="outline" action="secondary">
-                <ButtonText>Back</ButtonText>
-              </Button>
-              <Button>
-                <ButtonText>Next</ButtonText>
-              </Button>
-            </ButtonGroup>
-          </PopoverFooter>
-        </PopoverContent>
-      </Popover>
-    </Box>
   );
 };
 
@@ -145,7 +89,6 @@ PopoverBasic.description = `Popover is a non-modal dialog that floats around a t
 export default PopoverBasic;
 
 export {
-  FigmaPopoverStory,
   PopoverBasic,
   Text,
   Popover,
@@ -156,23 +99,7 @@ export {
   PopoverFooter,
   PopoverCloseButton,
   Button,
-  ButtonGroup,
   ButtonText,
-  CloseIcon,
   Box,
-  Heading,
-  Icon,
-  Pressable,
-  HStack,
-  VStack,
-  Avatar,
-  AvatarFallbackText,
-  CircleIcon,
-  AddIcon,
-  Center,
-  PhoneIcon,
-  Clock3Icon,
-  MailIcon,
   useState,
-  PopoverArrow,
 };